fix: remove spurious (short) casts in Size setters (#1791)
The Size fields are int but setColumns()/setRows() silently truncated
values via (short) casts since the class was created in 2016. This
served no purpose — the truncated value was stored back into an int
field, so there was no memory saving, only silent data loss for values
above 32767.

+++ terminal/src/main/java/org/jline/terminal/Size.java
@@ -104,7 +104,7 @@ public class Size {
      * @see #getColumns()
      */
     public void setColumns(int columns) {
-        cols = (short) columns;
+        cols = columns;
     }
 
     /**
@@ -132,7 +132,7 @@ public class Size {
      * @see #getRows()
      */
     public void setRows(int rows) {
-        this.rows = (short) rows;
+        this.rows = rows;
     }
 
     /**
